+++ pkg/api/api.go
@@ -5,9 +5,11 @@ import (
 	"errors"
 	"fmt"
 	neturl "net/url"
+	"strings"
 	"time"
 
 	"github.com/gin-gonic/gin"
+	"github.com/tuvistavie/securerandom"
 
 	"github.com/sosedoff/pgweb/pkg/bookmarks"
 	"github.com/sosedoff/pgweb/pkg/client"
@@ -68,6 +70,51 @@ func GetSessions(c *gin.Context) {
 	c.JSON(200, map[string]int{"sessions": len(DbSessions)})
 }
 
+func ConnectWithBackend(c *gin.Context) {
+	// Setup a new backend client
+	backend := Backend{
+		Endpoint:    command.Opts.ConnectBackend,
+		Token:       command.Opts.ConnectToken,
+		PassHeaders: command.Opts.ConnectHeaders,
+	}
+
+	// Fetch connection credentials
+	cred, err := backend.FetchCredential(c.Param("resource"), c)
+	if err != nil {
+		c.JSON(400, Error{err.Error()})
+		return
+	}
+
+	// Make the new session
+	sessionId, err := securerandom.Uuid()
+	if err != nil {
+		c.JSON(400, Error{err.Error()})
+		return
+	}
+	c.Request.Header.Add("x-session-id", sessionId)
+
+	// Connect to the database
+	cl, err := client.NewFromUrl(cred.DatabaseUrl, nil)
+	if err != nil {
+		c.JSON(400, Error{err.Error()})
+		return
+	}
+	cl.External = true
+
+	// Finalize session seetup
+	_, err = cl.Info()
+	if err == nil {
+		err = setClient(c, cl)
+	}
+	if err != nil {
+		cl.Close()
+		c.JSON(400, Error{err.Error()})
+		return
+	}
+
+	c.Redirect(301, fmt.Sprintf("/%s?session=%s", command.Opts.Prefix, sessionId))
+}
+
 func Connect(c *gin.Context) {
 	if command.Opts.LockSession {
 		c.JSON(400, Error{"Session is locked"})
@@ -109,11 +156,11 @@ func Connect(c *gin.Context) {
 	info, err := cl.Info()
 	if err == nil {
 		err = setClient(c, cl)
-		if err != nil {
-			cl.Close()
-			c.JSON(400, Error{err.Error()})
-			return
-		}
+	}
+	if err != nil {
+		cl.Close()
+		c.JSON(400, Error{err.Error()})
+		return
 	}
 
 	c.JSON(200, info.Format()[0])
@@ -140,6 +187,12 @@ func SwitchDb(c *gin.Context) {
 		return
 	}
 
+	// Do not allow switching databases for connections from third-party backends
+	if conn.External {
+		c.JSON(400, Error{"Session is locked"})
+		return
+	}
+
 	currentUrl, err := neturl.Parse(conn.ConnectionString)
 	if err != nil {
 		c.JSON(400, Error{"Unable to parse current connection string"})
@@ -163,11 +216,11 @@ func SwitchDb(c *gin.Context) {
 	info, err := cl.Info()
 	if err == nil {
 		err = setClient(c, cl)
-		if err != nil {
-			cl.Close()
-			c.JSON(400, Error{err.Error()})
-			return
-		}
+	}
+	if err != nil {
+		cl.Close()
+		c.JSON(400, Error{err.Error()})
+		return
 	}
 
 	conn.Close()
@@ -198,6 +251,12 @@ func Disconnect(c *gin.Context) {
 }
 
 func GetDatabases(c *gin.Context) {
+	conn := DB(c)
+	if conn.External {
+		c.JSON(403, Error{"Not permitted"})
+		return
+	}
+
 	names, err := DB(c).Databases()
 	serveResult(names, err, c)
 }
@@ -397,4 +456,40 @@ func GetInfo(c *gin.Context) {
 	}
 
 	c.JSON(200, info)
+}
+
+// Export database or table data
+func DataExport(c *gin.Context) {
+	db := DB(c)
+
+	info, err := db.Info()
+	if err != nil {
+		c.JSON(400, Error{err.Error()})
+		return
+	}
+
+	dump := client.Dump{
+		Table: strings.TrimSpace(c.Request.FormValue("table")),
+	}
+
+	// If pg_dump is not available the following code will not show an error in browser.
+	// This is due to the headers being written first.
+	if !dump.CanExport() {
+		c.JSON(400, Error{"pg_dump is not found"})
+		return
+	}
+
+	formattedInfo := info.Format()[0]
+	filename := formattedInfo["current_database"].(string)
+	if dump.Table != "" {
+		filename = filename + "_" + dump.Table
+	}
+
+	attachment := fmt.Sprintf(`attachment; filename="%s.sql.gz"`, filename)
+	c.Header("Content-Disposition", attachment)
+
+	err = dump.Export(db.ConnectionString, c.Writer)
+	if err != nil {
+		c.JSON(400, Error{err.Error()})
+	}
 }