+++ pkg/bookmarks/bookmarks.go
@@ -14,18 +14,18 @@ import (
 )
 
 type Bookmark struct {
-	Url      string         `json:"url"`      // Postgres connection URL
-	Host     string         `json:"host"`     // Server hostname
-	Port     int            `json:"port"`     // Server port
-	User     string         `json:"user"`     // Database user
-	Password string         `json:"password"` // User password
-	Database string         `json:"database"` // Database name
-	Ssl      string         `json:"ssl"`      // Connection SSL mode
-	Ssh      shared.SSHInfo `json:"ssh"`      // SSH tunnel config
+	Url      string          `json:"url"`      // Postgres connection URL
+	Host     string          `json:"host"`     // Server hostname
+	Port     int             `json:"port"`     // Server port
+	User     string          `json:"user"`     // Database user
+	Password string          `json:"password"` // User password
+	Database string          `json:"database"` // Database name
+	Ssl      string          `json:"ssl"`      // Connection SSL mode
+	Ssh      *shared.SSHInfo `json:"ssh"`      // SSH tunnel config
 }
 
 func (b Bookmark) SSHInfoIsEmpty() bool {
-	return b.Ssh.User == "" && b.Ssh.Host == "" && b.Ssh.Port == ""
+	return b.Ssh == nil || b.Ssh.User == "" && b.Ssh.Host == "" && b.Ssh.Port == ""
 }
 
 func (b Bookmark) ConvertToOptions() command.Options {
@@ -52,6 +52,27 @@ func readServerConfig(path string) (Bookmark, error) {
 
 	if bookmark.Port == 0 {
 		bookmark.Port = 5432
+	}
+
+	// List of all supported postgres modes
+	modes := []string{"disable", "allow", "prefer", "require", "verify-ca", "verify-full"}
+	valid := false
+
+	for _, mode := range modes {
+		if bookmark.Ssl == mode {
+			valid = true
+			break
+		}
+	}
+
+	// Fall back to a default mode if mode is not set or invalid
+	// Typical typo: ssl mode set to "disabled"
+	if bookmark.Ssl == "" || !valid {
+		bookmark.Ssl = "disable"
+	}
+
+	if bookmark.Ssh != nil && bookmark.Ssh.Port == "" {
+		bookmark.Ssh.Port = "22"
 	}
 
 	return bookmark, err