+++ pkg/client/client.go
@@ -4,7 +4,9 @@ import (
 	"fmt"
 	neturl "net/url"
 	"reflect"
+	"regexp"
 	"strings"
+	"time"
 
 	_ "github.com/lib/pq"
 
@@ -16,9 +18,21 @@ import (
 	"github.com/sosedoff/pgweb/pkg/statements"
 )
 
+var (
+	postgresSignature = regexp.MustCompile(`(?i)postgresql ([\d\.]+)\s`)
+	postgresType      = "PostgreSQL"
+
+	cockroachSignature = regexp.MustCompile(`(?i)cockroachdb ccl v([\d\.]+)\s`)
+	cockroachType      = "CockroachDB"
+)
+
 type Client struct {
 	db               *sqlx.DB
 	tunnel           *Tunnel
+	serverVersion    string
+	serverType       string
+	lastQueryTime    time.Time
+	External         bool
 	History          []history.Record `json:"history"`
 	ConnectionString string           `json:"connection_string"`
 }
@@ -62,6 +76,7 @@ func New() (*Client, error) {
 		History:          history.New(),
 	}
 
+	client.setServerVersion()
 	return &client, nil
 }
 
@@ -69,6 +84,9 @@ func NewFromUrl(url string, sshInfo *shared.SSHInfo) (*Client, error) {
 	var tunnel *Tunnel
 
 	if sshInfo != nil {
+		if command.Opts.DisableSSH {
+			return nil, fmt.Errorf("ssh connections are disabled")
+		}
 		if command.Opts.Debug {
 			fmt.Println("Opening SSH tunnel for:", sshInfo)
 		}
@@ -101,6 +119,11 @@ func NewFromUrl(url string, sshInfo *shared.SSHInfo) (*Client, error) {
 		fmt.Println("Creating a new client for:", url)
 	}
 
+	uri, err := neturl.Parse(url)
+	if err == nil && uri.Path == "" {
+		return nil, fmt.Errorf("Database name is not provided")
+	}
+
 	db, err := sqlx.Open("postgres", url)
 	if err != nil {
 		return nil, err
@@ -113,7 +136,33 @@ func NewFromUrl(url string, sshInfo *shared.SSHInfo) (*Client, error) {
 		History:          history.New(),
 	}
 
+	client.setServerVersion()
 	return &client, nil
+}
+
+func (client *Client) setServerVersion() {
+	res, err := client.query("SELECT version()")
+	if err != nil || len(res.Rows) < 1 {
+		return
+	}
+
+	version := res.Rows[0][0].(string)
+
+	// Detect postgresql
+	matches := postgresSignature.FindAllStringSubmatch(version, 1)
+	if len(matches) > 0 {
+		client.serverType = postgresType
+		client.serverVersion = matches[0][1]
+		return
+	}
+
+	// Detect cockroachdb
+	matches = cockroachSignature.FindAllStringSubmatch(version, 1)
+	if len(matches) > 0 {
+		client.serverType = cockroachType
+		client.serverVersion = matches[0][1]
+		return
+	}
 }
 
 func (client *Client) Test() error {
@@ -172,19 +221,47 @@ func (client *Client) TableRows(table string, opts RowsOptions) (*Result, error)
 	return client.query(sql)
 }
 
+func (client *Client) EstimatedTableRowsCount(table string, opts RowsOptions) (*Result, error) {
+	schema, table := getSchemaAndTable(table)
+	sql := fmt.Sprintf(`SELECT reltuples FROM pg_class WHERE oid = '%s.%s'::regclass;`, schema, table)
+
+	result, err := client.query(sql)
+	if err != nil {
+		return nil, err
+	}
+	// float64 to int64 conversion
+	estimatedRowsCount := result.Rows[0][0].(float64)
+	result.Rows[0] = Row{int64(estimatedRowsCount)}
+
+	return result, nil
+}
+
 func (client *Client) TableRowsCount(table string, opts RowsOptions) (*Result, error) {
 	schema, table := getSchemaAndTable(table)
 	sql := fmt.Sprintf(`SELECT COUNT(1) FROM "%s"."%s"`, schema, table)
 
 	if opts.Where != "" {
 		sql += fmt.Sprintf(" WHERE %s", opts.Where)
+	} else if client.serverType == postgresType {
+		tableInfo, err := client.TableInfo(table)
+		if err != nil {
+			return nil, err
+		}
+		estimatedRowsCount := tableInfo.Rows[0][3].(float64)
+		if estimatedRowsCount > 100000 {
+			return client.EstimatedTableRowsCount(table, opts)
+		}
 	}
 
 	return client.query(sql)
 }
 
 func (client *Client) TableInfo(table string) (*Result, error) {
-	return client.query(statements.TableInfo, table)
+	if client.serverType == cockroachType {
+		return client.query(statements.TableInfoCockroach)
+	}
+	schema, table := getSchemaAndTable(table)
+	return client.query(statements.TableInfo, fmt.Sprintf(`"%s"."%s"`, schema, table))
 }
 
 func (client *Client) TableIndexes(table string) (*Result, error) {
@@ -211,7 +288,17 @@ func (client *Client) TableConstraints(table string) (*Result, error) {
 
 // Returns all active queriers on the server
 func (client *Client) Activity() (*Result, error) {
-	return client.query(statements.Activity)
+	if client.serverType == cockroachType {
+		return client.query("SHOW QUERIES")
+	}
+
+	version := getMajorMinorVersion(client.serverVersion)
+	query := statements.Activity[version]
+	if query == "" {
+		query = statements.Activity["default"]
+	}
+
+	return client.query(query)
 }
 
 func (client *Client) Query(query string) (*Result, error) {
@@ -239,7 +326,16 @@ func (client *Client) SetReadOnlyMode() error {
 	return nil
 }
 
+func (client *Client) ServerVersion() string {
+	return fmt.Sprintf("%s %s", client.serverType, client.serverVersion)
+}
+
 func (client *Client) query(query string, args ...interface{}) (*Result, error) {
+	// Update the last usage time
+	defer func() {
+		client.lastQueryTime = time.Now().UTC()
+	}()
+
 	// We're going to force-set transaction mode on every query.
 	// This is needed so that default mode could not be changed by user.
 	if command.Opts.ReadOnly {
@@ -328,6 +424,16 @@ func (client *Client) Close() error {
 	}
 
 	return nil
+}
+
+func (client *Client) IsIdle() bool {
+	mins := int(time.Since(client.lastQueryTime).Minutes())
+
+	if command.Opts.ConnectionIdleTimeout > 0 {
+		return mins >= command.Opts.ConnectionIdleTimeout
+	}
+
+	return false
 }
 
 // Fetch all rows as strings for a single column