+++ pkg/client/result.go
@@ -8,6 +8,8 @@ import (
 	"reflect"
 	"strconv"
 	"time"
+
+	"github.com/sosedoff/pgweb/pkg/command"
 )
 
 type Row []interface{}
@@ -108,7 +110,14 @@ func (res *Result) CSV() []byte {
 }
 
 func (res *Result) JSON() []byte {
-	data, _ := json.Marshal(res.Format())
+	var data []byte
+
+	if command.Opts.DisablePrettyJson {
+		data, _ = json.Marshal(res.Format())
+	} else {
+		data, _ = json.MarshalIndent(res.Format(), "", " ")
+	}
+
 	return data
 }